+++ app/Http/Controllers/Admin/ResourceController.php
@@ -77,7 +77,10 @@ class ResourceController extends Controller
     //软删除移除资源
     public function remove(Request $request, Resource $resource)
     { 
-        //此处应该做一个判断，资源如果被使用，则禁止删除
+        if($resource->chapter()->count()>0){
+            alert('删除失败，该资源已绑定章节','danger');
+            return back();
+        }
         $resource->delete();
         alert('删除成功');
         return redirect()->route('admin.resource');
